Move file to the right place
