+++ src/VirtoCommerce.DescriptionExportImportModule.Core/Services/IDescriptionDataExporter.cs
@@ -1,8 +1,9 @@
 using System;
 using System.Threading.Tasks;
+using VirtoCommerce.DescriptionExportImportModule.Core.Models;
 using VirtoCommerce.Platform.Core.Common;
 
-namespace VirtoCommerce.DescriptionExportImportModule.Core.Models
+namespace VirtoCommerce.DescriptionExportImportModule.Core.Services
 {
     public interface IDescriptionDataExporter
     {
